+++ website/common/script/libs/inAppRewards.js
@@ -1,6 +1,7 @@
 import getItemInfo from './getItemInfo';
 import shops from './shops';
 import getOfficialPinnedItems from './getOfficialPinnedItems';
+import getDebuffSpellItems from './getDebuffSpellItems';
 import compactArray from 'lodash/compact';
 
 import getItemByPathAndType from './getItemByPathAndType';
@@ -50,6 +51,10 @@ module.exports = function getPinnedItems (user) {
 
       return getItemInfo(user, type, item, officialPinnedItems);
     });
+
+  const debuffSpellItems = getDebuffSpellItems(user);
+
+  items = items.concat(debuffSpellItems);
 
   shops.checkMarketGearLocked(user, items);
